REFACTOR: refactored cases into separate module
